--- a/barretenberg/cpp/scripts/compare_branch_vs_baseline.sh
+++ b/barretenberg/cpp/scripts/compare_branch_vs_baseline.sh
@@ -5,10 +5,12 @@
 # it is up to date with local master, and run the script.
 
 # Specify the benchmark suite and the "baseline" branch against which to compare
-BENCH_TARGET=${1:?"Please provide the name of a benchmark target."}
+BENCHMARK=${1:?"Please provide the name of a benchmark target."}
+COMMAND=${2:-./$BENCHMARK}
 BASELINE_BRANCH="master"
 
-echo -e "\nComparing $BENCH_TARGET between $BASELINE_BRANCH and current branch:"
+
+echo -e "\nComparing $BENCHMARK between $BASELINE_BRANCH and current branch:"
 # Set some directories
 BASE_DIR="$HOME/aztec-packages/barretenberg/cpp"
 BUILD_DIR="$BASE_DIR/build-bench" # matches build dir specified in bench preset
@@ -24,24 +26,24 @@ cd $BASE_DIR
 mkdir $BENCH_RESULTS_DIR
 
 # Build and run bench in current branch
-echo -e "\nConfiguring and building $BENCH_TARGET in current feature branch..\n"
-rm -rf $BUILD_DIR
-cmake --preset bench > /dev/null && cmake --build --preset bench --target $BENCH_TARGET
+echo -e "\nConfiguring and building $BENCHMARK in current feature branch..\n"
+# rm -rf $BUILD_DIR
+cmake --preset bench > /dev/null && cmake --build --preset bench --target $BENCHMARK
 cd build-bench
 BRANCH_RESULTS="$BENCH_RESULTS_DIR/results_branch.json"
-echo -e "\nRunning $BENCH_TARGET in feature branch.."
-bin/$BENCH_TARGET --benchmark_format=json > $BRANCH_RESULTS
+echo -e "\nRunning $COMMAND in feature branch.."
+bin/$COMMAND --benchmark_format=json > $BRANCH_RESULTS
 
 # Checkout baseline branch, run benchmarks, save results in json format
-echo -e "\nConfiguring and building $BENCH_TARGET in $BASELINE_BRANCH branch..\n"
+echo -e "\nConfiguring and building $BENCHMARK in $BASELINE_BRANCH branch..\n"
 git checkout master > /dev/null
 cd $BASE_DIR
 rm -rf $BUILD_DIR
-cmake --preset bench > /dev/null && cmake --build --preset bench --target $BENCH_TARGET
+cmake --preset bench > /dev/null && cmake --build --preset bench --target $BENCHMARK
 cd build-bench
 BASELINE_RESULTS="$BENCH_RESULTS_DIR/results_baseline.json"
-echo -e "\nRunning $BENCH_TARGET in master.."
-bin/$BENCH_TARGET --benchmark_format=json > $BASELINE_RESULTS
+echo -e "\nRunning $COMMAND in master.."
+bin/$COMMAND --benchmark_format=json > $BASELINE_RESULTS
 
 # Call compare.py on the results (json) to get high level statistics.
 # See docs at https://github.com/google/benchmark/blob/main/docs/tools.md for more details.