--- a/barretenberg/cpp/scripts/compare_branch_vs_baseline_remote.sh
+++ b/barretenberg/cpp/scripts/compare_branch_vs_baseline_remote.sh
@@ -37,6 +37,7 @@ echo -e "\nRunning benchmark in feature branch.."
                                             --benchmark_out_format=json"\
                               $PRESET
                               $BUILD_DIR
+scp $BB_SSH_KEY $BB_SSH_INSTANCE:$BB_SSH_CPP_PATH/build/results_after.json $BUILD_DIR/
 
 # Configure and build benchmark in $BASELINE branch
 echo -e "\nConfiguring and building $BENCHMARK in $BASELINE_BRANCH...\n"
@@ -44,7 +45,7 @@ git checkout $BASELINE_BRANCH
 cmake --preset $PRESET
 cmake --build --preset $PRESET --target $BENCHMARK
 
-zs# Run bench in current branch
+# Run bench in current branch
 echo -e "\nRunning benchmark in feature branch.."
 ./scripts/benchmark_remote.sh $BENCHMARK\
                               "./$BENCHMARK --benchmark_filter=$FILTER\
@@ -52,6 +53,7 @@ echo -e "\nRunning benchmark in feature branch.."
                                             --benchmark_out_format=json"\
                               $PRESET
                               $BUILD_DIR
+scp $BB_SSH_KEY $BB_SSH_INSTANCE:$BB_SSH_CPP_PATH/build/results_before.json $BUILD_DIR/
 
 # Call compare.py on the results (json) to get high level statistics.
 # See docs at https://github.com/google/benchmark/blob/main/docs/tools.md for more details.