… barretenberg/cpp/src/barretenberg/eccvm/msm_builder.hpp